Loaf/market prices

+++ contracts/market/src/constants.cairo
@@ -1,11 +1,11 @@
 // number of items per level per survivor
-const NUMBER_OF_ITEMS_PER_LEVEL: u32 = 20;
+const NUMBER_OF_ITEMS_PER_LEVEL: u64 = 20;
 
 // randomised offset for each survivor
-const OFFSET: u32 = 9636;
+const OFFSET: u64 = 9636;
 
 // number of loot items
-const NUM_LOOT_ITEMS: u32 = 101;
+const NUM_LOOT_ITEMS: u64 = 101;
 
 
-const TIER_PRICE: u8 = 3;
+const TIER_PRICE: u16 = 3;
